+++ src/config.ts
@@ -42,6 +42,7 @@ export type ConfigType = {
     // Trending tags
     excessiveTags: number;
     excessiveTagsPenalty: number;
+    invalidTrendingTags: string[];
     minTrendingTagTootsForPenalty: number,
     numTootsPerTrendingTag: number;
     numDaysToCountTrendingTagData: number;
@@ -112,6 +113,9 @@ export const Config: ConfigType = {
     // Trending tags and links
     excessiveTags: 25,                   // Toots with more than this many tags will be penalized
     excessiveTagsPenalty: 0.1,           // Multiplier to penalize toots with excessive tags
+    invalidTrendingTags: [               // Tags that are too generic to be considered trending
+        "news",
+    ],
     minTrendingTagTootsForPenalty: 9,    // Minimum number of toots with a trending tag before DiversityFeedScorer applies a penalty
     numDaysToCountTrendingTagData: 3,    // Look at this many days of user counts when assessing trending tags
     numTootsPerTrendingTag: 15,          // How many toots to pull for each trending tag